[GOHW-3] Домашка с gRPC

See merge request alexanderkoltygin/homework!3

+++ cart/internal/cart/model/item.go
@@ -5,6 +5,6 @@ type SkuID int64
 type Item struct {
 	SKU   SkuID
 	Name  string
-	Count uint16
+	Count int64
 	Price uint32
 }
